Don't attempt to delete the file unless it exists (fixes error when rescueing a download error that never created a file)

+++ lib/content/entry_importer/mods_file.rb
@@ -37,7 +37,7 @@ class Content::EntryImporter::ModsFile
       doc = Nokogiri::XML(open(file_path))
       raise Content::EntryImporter::ModsFile::DownloadError unless doc.root.name == "mods"
     rescue
-      File.delete(file_path)
+      File.delete(file_path) if File.exists?(file_path)
 
       if (retry_attempts -= 1) > 0
         sleep 10
